refactor!: split Perimeter implementations and improve related types
Introduce a PerimeterFunction type to precisely document the function accepted as perimeter. This
better guides implementers of perimeter. This required to update the code calling the perimeter
function (mainly to manage `null` values).

Split Perimeter implementations. The "Perimeter" file was too large, the change also prepare better
tree-shaking in the future for people that don't want to use all default perimeter implementation.
In addition, fix the Hexagon perimeter to match the direction style value of the underlying hexagon
shape. The "Perimeter" story has been updated to use an orthogonal direction and show the fix.

`Perimeter` is no longer a class but a value object. It only contains static methods and properties.

Add various JSDoc improvements.

BREAKING CHANGES
`Perimeter` is no longer a class but a value object. This only impact users that had extended the
`Perimeter` class. Regular user that define perimeter style properties using function provided by
`maxGraph` are not impacted by this change.

CellState.perimeter no longer accept `Function`, but only the `PerimeterFunction`.
  - Passing an arbitrary `Function` was incorrect and always failed at runtime.
  - This change should not impact people using working implementation of perimeter function
  (including these provided by maxGraph) as they already have the right signature. Implementers of
  custom perimeter in TypeScript may have to slightly update their perimeter function declaration.
